+++ crates/zizmor/src/audit/ref_confusion.rs
@@ -12,7 +12,7 @@ use github_actions_models::common::{RepositoryUses, Uses};
 use super::{Audit, AuditLoadError, Job, audit_meta};
 use crate::finding::Finding;
 use crate::finding::location::Locatable as _;
-use crate::models::{CompositeStep, StepCommon};
+use crate::models::{StepCommon, action::CompositeStep};
 use crate::{
     finding::{Confidence, Severity},
     github_api,
@@ -70,7 +70,7 @@ impl Audit for RefConfusion {
 
     fn audit_workflow<'doc>(
         &self,
-        workflow: &'doc crate::models::Workflow,
+        workflow: &'doc crate::models::workflow::Workflow,
     ) -> anyhow::Result<Vec<crate::finding::Finding<'doc>>> {
         let mut findings = vec![];